Moved Picture widget code to its own file
